Detect references that are actually roll table entries.
There are lots of roll tables, and they contains entries like "1-5" or "9-10"
which look like references to areas, but are actually just indicating the range
of results on which to use the corresponding entry in the table.

I first tried using the PyMuPDF table detection code, but it was way too slow.
It made us take approximately 20x the runtime we took before. So instead I've
used the approach in this commit:

1. While scaning for links, keep track of all words that could be entries in a
   roll table. This includes individual numbers, and two hyphen-separated
   numbers where the second is greater than the first.
2. Scan through this list of possible roll table entries, and find chains. For
   two entries to be adjacent in a chain, they must be directly adjacent
   numerically, have approximately equal x-coordinates, and have increasing
   y-coordinates.
3. For every such chain, take every entry which could be a link target and add
   its centre to a list of excluded points
4. Before emitting a link, check if the word's rectangle contains one of the
   excluded points, and if so, drop it.

This is very fast and adds very little to our runtime.